chore: add code syntax test case
