--- plugins/code-syntax-highlight/src/__test__/jest-setup.ts
@@ -1,3 +1,0 @@
-import '@testing-library/jest-dom';
-
-Element.prototype.scrollIntoView = jest.fn();
